types: move sidecar money fields to Decimal

+++ sidecar/app/schemas.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, Field
@@ -12,23 +13,23 @@ class ReserveRequest(BaseModel):
     trace_id: str = Field(..., min_length=1, max_length=255)
     idempotency_key: str = Field(..., min_length=1, max_length=255)
     model: str = Field(..., min_length=1, max_length=255)
-    estimated_cost: float = Field(..., ge=0)
+    estimated_cost: Decimal = Field(..., ge=0)
 
 
 class ReserveResponse(BaseModel):
     idempotency_key: str
     status: str
-    reserved_amount: float
+    reserved_amount: Decimal
     expires_in_seconds: int
 
 
 class SettleRequest(BaseModel):
     idempotency_key: str = Field(..., min_length=1, max_length=255)
-    actual_cost: float = Field(..., ge=0)
+    actual_cost: Decimal = Field(..., ge=0)
     provider_request_id: Optional[str] = Field(default=None, max_length=255)
 
 
 class SettleResponse(BaseModel):
     idempotency_key: str
     status: str
-    actual_amount: float
+    actual_amount: Decimal
